+++ Changes.md
@@ -10,6 +10,7 @@ Version 3.11.1.5 - TBR
 4. Fix 'Course overview' dropdown menu item 'All (except removed from view)' icon.
 5. Update FontAwesome from 5.12.2 to 5.12.4.
 6. Add FontAwesome 6.0.0-beta.
+7. Add Morph, Quartz, Vapor and Zephyr bootswatches.
 
 Version 3.11.1.4 - 08/10/21
 ---------------------------